requirements.txt
sanic~=25.3.0
sanic-ext~=24.12.0
httpx~=0.28.1
psutil~=7.0.0
pika~=1.3.2
redis~=6.2.0
python-binance
dotenv~=0.9.9
aio-pika~=9.5.5
web3~=7.12.0
gspread~=6.2.1
pandas~=2.3.0
feedparser
celery~=5.5.3
sqlalchemy~=2.0.41
psycopg2-binary
python-dotenv~=1.1.0

eth-typing~=5.2.1
eth-utils~=5.3.0
hexbytes~=1.3.1
idna~=3.10
pyunormalize~=16.0.0
h11~=0.16.0
pip~=25.0.1
attrs~=25.3.0
rlp~=4.1.0
rsa~=4.9.1
pyasn1~=0.6.1
toolz~=1.0.0
pytz~=2025.2
eth-account~=0.13.7
eth_abi~=5.2.0
requests~=2.32.4
aiohttp~=3.12.12
typing_extensions~=4.14.0
pydantic~=2.11.6
websockets~=15.0.1
eth-keys~=0.7.0
PyYAML~=6.0.2
multidict~=6.4.4
propcache~=0.3.2
sniffio~=1.3.1
click~=8.2.1
httpcore~=1.0.9
numpy~=2.3.0
charset-normalizer~=3.4.2
setuptools~=80.9.0
pamqp~=3.3.0
regex~=2024.11.6
sanic-routing~=23.12.0
ujson~=5.10.0
html5tagger~=1.3.0
tracerite~=1.1.1
httptools~=0.6.4
uvloop~=0.21.0
aiofiles~=24.1.0
yarl~=1.20.1
aiormq~=6.8.1
pycryptodome~=3.23.0
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.4.26
urllib3~=2.4.0
packaging~=25.0
python-dateutil~=2.9.0.post0
aiosignal~=1.3.2
frozenlist~=1.7.0
aiohappyeyeballs~=2.6.1
dateparser~=1.2.1
cytoolz~=1.0.1
parsimonious~=0.10.0
google-auth-oauthlib~=1.2.2
tzlocal~=5.3.1
exceptiongroup~=1.3.0
bitarray~=3.4.2
six~=1.17.0
eth-hash~=0.7.1
anyio~=4.9.0
oauthlib~=3.2.2
pydantic_core~=2.33.2
annotated-types~=0.7.0
sgmllib3k~=1.0.0
eth-keyfile~=0.8.1
eth-rlp~=2.2.0
ckzg~=2.1.1
requests-oauthlib~=2.0.0